pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anki-add-ons"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
  "anki>=25.9.2",
  "aqt>=25.9.2",
  "python-dotenv>=1.0.0"
]

[tool.hatch.envs.default]
skip-install = true
dependencies = [
  "black",
  "ruff",
  "mypy"
]

[tool.hatch.envs.default.scripts]
build = "python scripts/build_all.py"
install-dev = [
  "python scripts/build_all.py",
  "python scripts/install_dev.py"
]
fmt = [
  "black .",
  "ruff check --fix ."
]
check = [
  "black --check .",
  "ruff check .",
  "mypy ."
]

[tool.uv]
package = false
